refactor(iroh-bytes): Remove the size argument from get_or_create (#2096)
refactor(iroh-bytes): Remove the size argument from get_or_create

We have *a* size, but we don't know if it is the correct size. At the
point where we call get_or_create it could be anything. So we must not
use this size for *anything* whatsoever, so we might as well omit it.

The purpose of the size is solely to make sense of the following data,
nothing else.

## Description

<!-- A summary of what this pull request achieves and a rough list of
changes. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.